Add EXTCODE* unit Tests (#7374)
Add unit tests to the EXTCODE* series operations.
Also, put all EVM operations tests in the proper package.
