+++ evm/src/main/java/org/hyperledger/besu/evm/MainnetEVMs.java
@@ -240,8 +240,8 @@ public class MainnetEVMs {
     registry.put(new CodeSizeOperation(gasCalculator));
     registry.put(new CodeCopyOperation(gasCalculator));
     registry.put(new GasPriceOperation(gasCalculator));
-    registry.put(new ExtCodeCopyOperation(gasCalculator));
-    registry.put(new ExtCodeSizeOperation(gasCalculator));
+    registry.put(new ExtCodeCopyOperation(gasCalculator, false));
+    registry.put(new ExtCodeSizeOperation(gasCalculator, false));
     registry.put(new BlockHashOperation(gasCalculator));
     registry.put(new CoinbaseOperation(gasCalculator));
     registry.put(new TimestampOperation(gasCalculator));
@@ -478,7 +478,7 @@ public class MainnetEVMs {
     registry.put(new SarOperation(gasCalculator));
     registry.put(new ShlOperation(gasCalculator));
     registry.put(new ShrOperation(gasCalculator));
-    registry.put(new ExtCodeHashOperation(gasCalculator));
+    registry.put(new ExtCodeHashOperation(gasCalculator, false));
   }
 
   /**
@@ -1113,6 +1113,11 @@ public class MainnetEVMs {
     registry.put(new DupNOperation(gasCalculator));
     registry.put(new SwapNOperation(gasCalculator));
     registry.put(new ExchangeOperation(gasCalculator));
+
+    // EIP-3540 EOF Aware EXTCODE* operations
+    registry.put(new ExtCodeCopyOperation(gasCalculator, true));
+    registry.put(new ExtCodeHashOperation(gasCalculator, true));
+    registry.put(new ExtCodeSizeOperation(gasCalculator, true));
 
     // EIP-4200 relative jump
     registry.put(new RelativeJumpOperation(gasCalculator));
