+++ evm/src/main/java/org/hyperledger/besu/evm/operation/ExtCodeCopyOperation.java
@@ -34,13 +34,26 @@ public class ExtCodeCopyOperation extends AbstractOperation {
   /** This is the "code" legacy contracts see when copying code from an EOF contract. */
   public static final Bytes EOF_REPLACEMENT_CODE = Bytes.fromHexString("0xef00");
 
+  private final boolean enableEIP3540;
+
   /**
    * Instantiates a new Ext code copy operation.
    *
    * @param gasCalculator the gas calculator
    */
   public ExtCodeCopyOperation(final GasCalculator gasCalculator) {
+    this(gasCalculator, false);
+  }
+
+  /**
+   * Instantiates a new Ext code copy operation.
+   *
+   * @param gasCalculator the gas calculator
+   * @param enableEIP3540 enable EIP-3540 semantics (don't copy EOF)
+   */
+  public ExtCodeCopyOperation(final GasCalculator gasCalculator, final boolean enableEIP3540) {
     super(0x3C, "EXTCODECOPY", 4, 0, gasCalculator);
+    this.enableEIP3540 = enableEIP3540;
   }
 
   /**
@@ -82,7 +95,10 @@ public class ExtCodeCopyOperation extends AbstractOperation {
     final Account account = frame.getWorldUpdater().get(address);
     final Bytes code = account != null ? account.getCode() : Bytes.EMPTY;
 
-    if (code.size() >= 2 && code.get(0) == EOFLayout.EOF_PREFIX_BYTE && code.get(1) == 0) {
+    if (enableEIP3540
+        && code.size() >= 2
+        && code.get(0) == EOFLayout.EOF_PREFIX_BYTE
+        && code.get(1) == 0) {
       frame.writeMemory(memOffset, sourceOffset, numBytes, EOF_REPLACEMENT_CODE);
     } else {
       frame.writeMemory(memOffset, sourceOffset, numBytes, code);
