+++ evm/src/main/java/org/hyperledger/besu/evm/operation/ExtCodeSizeOperation.java
@@ -32,13 +32,26 @@ public class ExtCodeSizeOperation extends AbstractOperation {
 
   static final Bytes EOF_SIZE = Bytes.of(2);
 
+  private final boolean enableEIP3540;
+
   /**
    * Instantiates a new Ext code size operation.
    *
    * @param gasCalculator the gas calculator
    */
   public ExtCodeSizeOperation(final GasCalculator gasCalculator) {
+    this(gasCalculator, false);
+  }
+
+  /**
+   * Instantiates a new Ext code size operation.
+   *
+   * @param gasCalculator the gas calculator
+   * @param enableEIP3540 enable EIP-3540 semantics (EOF is size 2)
+   */
+  public ExtCodeSizeOperation(final GasCalculator gasCalculator, final boolean enableEIP3540) {
     super(0x3B, "EXTCODESIZE", 1, 1, gasCalculator);
+    this.enableEIP3540 = enableEIP3540;
   }
 
   /**
@@ -70,7 +83,10 @@ public class ExtCodeSizeOperation extends AbstractOperation {
           codeSize = Bytes.EMPTY;
         } else {
           final Bytes code = account.getCode();
-          if (code.size() >= 2 && code.get(0) == EOFLayout.EOF_PREFIX_BYTE && code.get(1) == 0) {
+          if (enableEIP3540
+              && code.size() >= 2
+              && code.get(0) == EOFLayout.EOF_PREFIX_BYTE
+              && code.get(1) == 0) {
             codeSize = EOF_SIZE;
           } else {
             codeSize = Words.intBytes(code.size());
