+++ evm/src/test/java/org/hyperledger/besu/evm/operation/ConstantinopleSStoreOperationGasCostTest.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.evm.operations;
+package org.hyperledger.besu.evm.operation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -23,50 +23,46 @@ import org.hyperledger.besu.evm.internal.EvmConfiguration;
 import org.hyperledger.besu.evm.testutils.TestCodeExecutor;
 
 import org.apache.tuweni.units.bigints.UInt256;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class LondonSStoreOperationGasCostTest {
+public class ConstantinopleSStoreOperationGasCostTest {
 
   public static Object[][] scenarios() {
-    // Tests specified in EIP-3539.
+    // Tests specified in EIP-1283.
     return new Object[][] {
-      {"0x60006000556000600055", 0, 212, 0},
-      {"0x60006000556001600055", 0, 20112, 0},
-      {"0x60016000556000600055", 0, 20112, 19900},
-      {"0x60016000556002600055", 0, 20112, 0},
-      {"0x60016000556001600055", 0, 20112, 0},
-      {"0x60006000556000600055", 1, 3012, 4800},
-      {"0x60006000556001600055", 1, 3012, 2800},
-      {"0x60006000556002600055", 1, 3012, 0},
-      {"0x60026000556000600055", 1, 3012, 4800},
-      {"0x60026000556003600055", 1, 3012, 0},
-      {"0x60026000556001600055", 1, 3012, 2800},
-      {"0x60026000556002600055", 1, 3012, 0},
-      {"0x60016000556000600055", 1, 3012, 4800},
-      {"0x60016000556002600055", 1, 3012, 0},
-      {"0x60016000556001600055", 1, 212, 0},
-      {"0x600160005560006000556001600055", 0, 40118, 19900},
-      {"0x600060005560016000556000600055", 1, 5918, 7600},
+      {"0x60006000556000600055", 0, 412, 0},
+      {"0x60006000556001600055", 0, 20212, 0},
+      {"0x60016000556000600055", 0, 20212, 19800},
+      {"0x60016000556002600055", 0, 20212, 0},
+      {"0x60016000556001600055", 0, 20212, 0},
+      {"0x60006000556000600055", 1, 5212, 15000},
+      {"0x60006000556001600055", 1, 5212, 4800},
+      {"0x60006000556002600055", 1, 5212, 0},
+      {"0x60026000556003600055", 1, 5212, 0},
+      {"0x60026000556001600055", 1, 5212, 4800},
+      {"0x60026000556002600055", 1, 5212, 0},
+      {"0x60016000556000600055", 1, 5212, 15000},
+      {"0x60016000556002600055", 1, 5212, 0},
+      {"0x60016000556001600055", 1, 412, 0},
+      {"0x600160005560006000556001600055", 0, 40218, 19800},
+      {"0x600060005560016000556000600055", 1, 10218, 19800},
+      {"0x60026000556000600055", 1, 5212, 15000},
     };
-  }
-
-  private TestCodeExecutor codeExecutor;
-
-  @BeforeEach
-  public void setUp() {
-    codeExecutor = new TestCodeExecutor(MainnetEVMs.london(EvmConfiguration.DEFAULT));
   }
 
   @ParameterizedTest
   @MethodSource("scenarios")
-  void shouldCalculateGasAccordingToEip3529(
+  void shouldCalculateGasAccordingToEip1283(
       final String code,
       final int originalValue,
       final int expectedGasUsed,
       final int expectedGasRefund) {
+
+    TestCodeExecutor codeExecutor =
+        new TestCodeExecutor(MainnetEVMs.constantinople(EvmConfiguration.DEFAULT));
+
     final long gasLimit = 1_000_000;
     final MessageFrame frame =
         codeExecutor.executeCode(
@@ -74,7 +70,8 @@ public class LondonSStoreOperationGasCostTest {
             gasLimit,
             account -> account.setStorageValue(UInt256.ZERO, UInt256.valueOf(originalValue)));
     assertThat(frame.getState()).isEqualTo(State.COMPLETED_SUCCESS);
-    assertThat(frame.getRemainingGas()).isEqualTo(gasLimit - (expectedGasUsed + 2100));
+    System.out.println(gasLimit - frame.getRemainingGas());
+    assertThat(frame.getRemainingGas()).isEqualTo(gasLimit - expectedGasUsed);
     assertThat(frame.getGasRefund()).isEqualTo(expectedGasRefund);
   }
