+++ evm/src/test/java/org/hyperledger/besu/evm/operation/Push0OperationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright ConsenSys AG.
+ * Copyright contributors to Hyperledger Besu.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -12,57 +12,39 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.evm.operations;
+package org.hyperledger.besu.evm.operation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import org.hyperledger.besu.datatypes.Wei;
+import org.hyperledger.besu.evm.code.CodeV0;
 import org.hyperledger.besu.evm.frame.BlockValues;
-import org.hyperledger.besu.evm.frame.ExceptionalHaltReason;
 import org.hyperledger.besu.evm.frame.MessageFrame;
 import org.hyperledger.besu.evm.gascalculator.BerlinGasCalculator;
 import org.hyperledger.besu.evm.gascalculator.GasCalculator;
-import org.hyperledger.besu.evm.internal.Words;
-import org.hyperledger.besu.evm.operation.BaseFeeOperation;
-import org.hyperledger.besu.evm.operation.Operation;
 import org.hyperledger.besu.evm.operation.Operation.OperationResult;
+import org.hyperledger.besu.evm.testutils.FakeBlockValues;
 
 import java.util.Optional;
 
-import org.apache.tuweni.bytes.Bytes32;
-import org.apache.tuweni.units.bigints.UInt256;
+import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
-class BaseFeeOperationTest {
+class Push0OperationTest {
+
   private final GasCalculator gasCalculator = new BerlinGasCalculator();
 
   @Test
-  void shouldReturnGasCost() {
+  void shouldPush0OntoStack() {
     final MessageFrame frame = createMessageFrame(100, Optional.of(Wei.of(5L)));
-    final Operation operation = new BaseFeeOperation(gasCalculator);
+    final Operation operation = new Push0Operation(gasCalculator);
     final OperationResult result = operation.execute(frame, null);
+    Mockito.verify(frame).pushStackItem(Bytes.EMPTY);
     assertThat(result.getGasCost()).isEqualTo(gasCalculator.getBaseTierGasCost());
     assertSuccessResult(result);
-  }
-
-  @Test
-  void shouldWriteBaseFeeToStack() {
-    final MessageFrame frame = createMessageFrame(100, Optional.of(Wei.of(5L)));
-    final Operation operation = new BaseFeeOperation(gasCalculator);
-    final OperationResult result = operation.execute(frame, null);
-    verify(frame).pushStackItem(UInt256.fromBytes(Bytes32.leftPad(Words.longBytes(5L))));
-    assertSuccessResult(result);
-  }
-
-  @Test
-  void shouldHaltIfNoBaseFeeInBlockHeader() {
-    final MessageFrame frame = createMessageFrame(100, Optional.empty());
-    final Operation operation = new BaseFeeOperation(gasCalculator);
-    final OperationResult result = operation.execute(frame, null);
-    assertExceptionalHalt(result, ExceptionalHaltReason.INVALID_OPERATION);
   }
 
   private void assertSuccessResult(final OperationResult result) {
@@ -70,18 +52,12 @@ class BaseFeeOperationTest {
     assertThat(result.getHaltReason()).isNull();
   }
 
-  private void assertExceptionalHalt(
-      final OperationResult result, final ExceptionalHaltReason reason) {
-    assertThat(result).isNotNull();
-    assertThat(result.getHaltReason()).isEqualTo(reason);
-  }
-
   private MessageFrame createMessageFrame(final long initialGas, final Optional<Wei> baseFee) {
     final MessageFrame frame = mock(MessageFrame.class);
     when(frame.getRemainingGas()).thenReturn(initialGas);
-    final BlockValues blockHeader = mock(BlockValues.class);
-    when(blockHeader.getBaseFee()).thenReturn(baseFee);
-    when(frame.getBlockValues()).thenReturn(blockHeader);
+    final BlockValues blockValues = new FakeBlockValues(baseFee);
+    when(frame.getBlockValues()).thenReturn(blockValues);
+    when(frame.getCode()).thenReturn(CodeV0.EMPTY_CODE);
     return frame;
   }
 }
